t/basic.t: Don't become broken when MAKEOPTS includes V=1
Some people set MAKEOPTS to include V=1, in order to coax autotools
based makefiles into being more verbose.

But when that ENV var is used to direct perl ebuilds for Dzil things,
well, it unintentionally uses "1" as the *next version*, and so breaks
all code that isn't expecting that to happen.

Given its not the job of tests to tell people their system is weird, it
just makes the scope "normal" ( that is, defaulting V to be unset
internally )

Bug: https://bugs.gentoo.org/737308
Bug: https://github.com/yanick/Dist-Zilla-Plugin-NextVersion-Semantic/pull/6